Move 4 commands to single header file
