--- Patterns/02-Command/01-Configure-Input-and-Actors/src/JumpCommand.hpp
@@ -1,18 +1,0 @@
-#pragma once
-
-#include "Command.hpp"
-
-namespace igpp::c02_01
-{
-
-// Snippet from the book
-class JumpCommand final : public Command
-{
-public:
-    void execute(GameActor& actor) final override
-    {
-        actor.jump();
-    }
-};
-
-} // namespace igpp::c02_01
