--- Patterns/02-Command/01-Configure-Input-and-Actors/src/StrongAttackCommand.hpp
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "Command.hpp"
-
-namespace igpp::c02_01
-{
-
-class StrongAttackCommand final : public Command
-{
-public:
-    void execute(GameActor& actor) final override
-    {
-        actor.strongAttack();
-    }
-};
-
-} // namespace igpp::c02_01
