--- Patterns/02-Command/01-Configure-Input-and-Actors/src/WeakAttackCommand.hpp
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "Command.hpp"
-
-namespace igpp::c02_01
-{
-
-class WeakAttackCommand final : public Command
-{
-public:
-    void execute(GameActor& actor) final override
-    {
-        actor.weakAttack();
-    }
-};
-
-} // namespace igpp::c02_01
